add terminate_js8call attribute and associated logic

+++ pyjs8call/appmonitor.py
@@ -47,6 +47,7 @@ class AppMonitor:
         headless (bool): Whether JS8Call is running headless using xvfb (see *start()*)
         args (list): Sequence of command line arguments to be passed to JS8Call, defaults to empty list
         restart (bool): Whether to restart the JS8Call application if it stops, defaults to False
+        terminate_js8call (bool): Whether to terminate the JS8Call application when stopping the app monitor
     '''
 
     def __init__(self, parent):
@@ -64,6 +65,7 @@ class AppMonitor:
         self.headless = False
         self.args = []
         self.restart = False
+        self.terminate_js8call = True
 
     def start(self, headless=False, args=None):
         ''' Start JS8Call application.
@@ -146,6 +148,9 @@ class AppMonitor:
         if not self.is_running():
             return
 
+        if not self.terminate_js8call:
+            return
+        
         self._js8call_proc.terminate()
 
         try:
@@ -302,8 +307,8 @@ class AppMonitor:
                 if self.restart:
                     # restart the whole system and reconnect
                     self._parent._client.restart()
-                
                 else:
+                    # exit if the js8call application is closed
                     self._parent._client.exit_tasks()
                     psutil.Process().terminate()
